refactor: 리뷰 타입 코드 분리
