+++ src/components/review/ReviewFakeForm.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-interface IProps {
+interface Props {
   setIsFormOpen: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-export default function ReviewFakeForm({ setIsFormOpen }: IProps) {
+export default function ReviewFakeForm({ setIsFormOpen }: Props) {
   return (
     <div className='flex mb-4' onClick={() => setIsFormOpen(true)}>
       <input
